requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.29.0
pgvector>=0.3.0
openai>=1.30.0
pydantic-settings>=2.2.1
stripe>=9.0.0
